Next major update of eslint rules and plugins
The rules and plugins for eslint have not been updated in some time.
This change adds new rules added to the core as well as the plugins
since the last major update. It also adds plugins for chai-expect and
node.

A breaking change is that the configuration are now located in the a
config directory instead of in the project root.